Fix loading bundle extension
